pyproject.toml
[project]
name = "alpha_sdk"
version = "0.5.3"
description = "Everything that turns Claude into Alpha"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.9.0",
    "asyncpg>=0.30.0",
    "claude-agent-sdk==0.1.36",
    "feedparser>=6.0.0",
    "html2text>=2024.2.26",
    "httpx>=0.28.0",
    "logfire[httpx]>=3.0.0",
    "pendulum>=3.0.0",
    "pgvector>=0.3.0",
    "psycopg[binary]>=3.0.0",
    "python-frontmatter>=1.0.0",
    "redis>=5.0.0",
    "pyyaml>=6.0.0",
    "datasets>=4.5.0",
    "Pillow>=10.0.0",
    "yt-dlp>=2024.0.0",
    # CLI dependencies
    "typer>=0.12.0",
    "rich>=13.0.0",
]

[project.scripts]
cortex = "alpha_sdk.cli.cortex:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/alpha_sdk"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
